chore: rename `reset-password` to `forgot-password` for clarity
